# backend/gateway/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /user/code:
        post:
            tags:
                - UserService
            description: 获取验证码
            operationId: UserService_GetVerificationCode
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/gateway.v1.GetVerificationCodeRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/gateway.v1.GetVerificationCodeResponse'
    /user/info:
        get:
            tags:
                - UserService
            description: 获取用户信息
            operationId: UserService_GetUserInfo
            parameters:
                - name: userId
                  in: query
                  description: '@gotags: json:"user_id,omitempty,string"'
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/gateway.v1.GetUserInfoResponse'
        put:
            tags:
                - UserService
            description: 更新用户信息
            operationId: UserService_UpdateUserInfo
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/gateway.v1.UpdateUserInfoRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/gateway.v1.UpdateUserInfoResponse'
    /user/login:
        post:
            tags:
                - UserService
            description: 登录
            operationId: UserService_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/gateway.v1.LoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/gateway.v1.LoginResponse'
    /user/register:
        post:
            tags:
                - UserService
            description: 注册
            operationId: UserService_Register
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/gateway.v1.RegisterRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/gateway.v1.RegisterResponse'
    /user/voucher:
        post:
            tags:
                - UserService
            operationId: UserService_BindUserVoucher
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/gateway.v1.BindUserVoucherRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/gateway.v1.BindUserVoucherResponse'
        delete:
            tags:
                - UserService
            operationId: UserService_UnbindUserVoucher
            parameters:
                - name: voucherType
                  in: query
                  schema:
                    type: integer
                    format: enum
                - name: voucher
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/gateway.v1.UnbindUserVoucherResponse'
...
components:
    schemas:
...
        gateway.v1.BindUserVoucherRequest:
            type: object
            properties:
                voucherType:
                    type: integer
                    format: enum
                voucher:
                    type: string
        gateway.v1.BindUserVoucherResponse:
            type: object
            properties: {}
...
        gateway.v1.GetUserInfoResponse:
            type: object
            properties:
                user:
                    $ref: '#/components/schemas/gateway.v1.User'
        gateway.v1.GetVerificationCodeRequest:
            type: object
            properties:
                mobile:
                    type: string
                email:
                    type: string
        gateway.v1.GetVerificationCodeResponse:
            type: object
            properties:
                codeId:
                    type: string
                    description: '@gotags: json:"code_id,omitempty,string"'
...
        gateway.v1.LoginRequest:
            type: object
            properties:
                mobile:
                    type: string
                email:
                    type: string
                password:
                    type: string
        gateway.v1.LoginResponse:
            type: object
            properties:
                token:
                    type: string
...
        gateway.v1.RegisterRequest:
            type: object
            properties:
                mobile:
                    type: string
                email:
                    type: string
                password:
                    type: string
                codeId:
                    type: string
                    description: '@gotags: json:"code_id,omitempty,string"'
                code:
                    type: string
        gateway.v1.RegisterResponse:
            type: object
            properties:
                userId:
                    type: string
                    description: '@gotags: json:"user_id,omitempty,string"'
...
        gateway.v1.UnbindUserVoucherResponse:
            type: object
            properties: {}
...
        gateway.v1.UpdateUserInfoRequest:
            type: object
            properties:
                userId:
                    type: string
                    description: '@gotags: json:"user_id,omitempty,string"'
                name:
                    type: string
                avatar:
                    type: string
                backgroundImage:
                    type: string
                signature:
                    type: string
        gateway.v1.UpdateUserInfoResponse:
            type: object
            properties: {}
        gateway.v1.User:
            type: object
            properties:
                id:
                    type: string
                    description: '@gotags: json:"id,omitempty,string"'
                name:
                    type: string
                avatar:
                    type: string
                backgroundImage:
                    type: string
                signature:
                    type: string
                mobile:
                    type: string
                email:
                    type: string
                followCount:
                    type: string
                    description: '@gotags: json:"followCount,omitempty,string"'
                followerCount:
                    type: string
                    description: '@gotags: json:"followerCount,omitempty,string"'
                totalFavorited:
                    type: string
                    description: '@gotags: json:"totalFavorited,omitempty,string"'
                workCount:
                    type: string
                    description: '@gotags: json:"workCount,omitempty,string"'
                favoriteCount:
                    type: string
                    description: '@gotags: json:"favoriteCount,omitempty,string"'
...